pyproject.toml
[project]
name = "guitts"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai-whisper>=20240930",
    "pyinstaller>=6.13.0",
]

